refactor: challenge handling
Split the big file handling challenges into smaller files for each function. Remove the big file
